netloc/xml: Split XML tool
The static XML management library is splitted in two parts:
    * one for building and exporting the topology XML file
    * one to provide a fall-through API in case libxml2 is
      not installed, or simply wrap the libxml2.

This is done in order to provide an independant XML writing tool
without having to include all the unused symbols needed to export the
topology. Moreover, this allow a code duplication removal between the
two methods (using or not libxml2).

Signed-off-by: Clement Foyer <clement.foyer@inria.fr>